move hcaptcha server code from an action to a lib
